Midterm_Challenge/requirements.txt
chainlit
langchain
langchain-openai
langchain-core
langgraph
langchain-community
tiktoken
qdrant-client
openai
python-dotenv
pymupdf 